AIRAVATA-2212 Stop checking for status change once detected
Previous code had a bug where it would continue checking for a status
change after a status change was detected. This resulted in the page
trying to reload while it was loading, prevent the page from ever
actually loading in some cases.

Also added checking the job status for changes as well.