AIRAVATA-2342 Reverting to using username in API
One difference between the Keycloak API and the WSO2 IS API is that
where WSO2 IS expects the username, Keycloak expects the Keycloak user
id.  This made for a mismatch in how to call the facade for WSO2 IS and
Keycloak.  However, now that I have a way to get the user id from the
username, I was able to revert some changes earlier so that now WSO2 IS
and Keycloak facades support the same methods.